refactor: move ban command to its own file
